+++ anypytools/macroutils.py
@@ -418,8 +418,46 @@ class Export(MacroCommand):
         for var, name in zip(self.var_list, self.name):
             if name is not None:
                 cmd.append(f'print "#### ANYPYTOOLS RENAME OUTPUT: {name}"')
-            cmd.append(f'classoperation {var} "Dump"')
+            cmd.append(f"print {var}")
         return "\n".join(cmd)
+
+
+class ExtendOutput(MacroCommand):
+    """Add a arbitrary values to to output. This value will be printed in
+       in the output so AnyPyTools will pick it up.
+
+    Parameters
+    ----------
+    var_name : str
+        Name of the variable in output.
+    value : any
+        The value to add to the output.
+
+    Examples:
+    ---------
+    >>> ExtendOutput('MyVar', 23.5)
+    print MyVar = 23.5;
+
+    >>> ExtendOutput('MyArray', np.array([1,2,3]))
+    print MyArray = {1,2,3};
+
+    >>> ExtendOutput('MyString', "Hello, World!")
+    print MyString = "Hello, World!";
+
+    """
+
+    def __init__(self, var_name: str, value):
+        self.var_name = var_name
+        self.value = value
+
+    def get_macro(self, index, **kwarg):
+        if isinstance(self.value, np.ndarray):
+            val_str = array2anyscript(self.value)
+        elif isinstance(self.value, str):
+            val_str = f"'{self.value}'"
+        else:
+            val_str = str(self.value)
+        return f'print "{self.var_name} = {val_str};"'
 
 
 class SaveDesign(MacroCommand):