+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # AnyPyTools Change Log
 
+## v1.19.0
+
+**Added:**
+* New `macro_commands.ExtendOutput()` helper macro that can add arbitrary values 
+  to the results output. This allows adding values that don't need to exist as 
+  variables in the AnyBody model. This is useful for adding extra metadata that will
+  appear in the results output when creating the macros.
+
+  ```python 
+  macro = [
+    mc.Load("MyModel.main.any"),
+    mc.ExtendOutput("SubjectID", "S001"),
+    mc.ExtendOutput("SubjectHeight", "1.8"),
+  ]
+  results = app.start_macro(macro)
+  assert results[0]["SubjectID"] == "S001"
+  assert results[0]["SubjectHeight"] == 1.8
+  ```
+* {meth}`results.to_dataframe()
+  <anypytools.tools.AnyPyProcessOutput.to_dataframe>` has a new argument
+  `exclude_task_info` which can exclude task information (variables starting
+  with 'task_') when exporting results to a dataframe
+
+
 ## v1.18
 
 **Added:**